Move session restore regressions to Swift Testing
