+++ cmuxTests/SessionPersistenceTests.swift
@@ -30,7 +30,6 @@ final class SessionPersistenceTests: XCTestCase {
         SessionSnapshotRepository(
             schemaVersion: SessionSnapshotSchema.currentVersion,
             bundleIdentifier: bundleIdentifier,
-            repairLoadedSnapshot: AppSessionSnapshot.repairLoadedSessionSnapshot,
             appSupportDirectory: appSupportDirectory
         )
     }
@@ -212,132 +211,6 @@ final class SessionPersistenceTests: XCTestCase {
         XCTAssertEqual(loaded?.windows.first?.display?.displayID, 42)
         let visibleFrame = try XCTUnwrap(loaded?.windows.first?.display?.visibleFrame)
         XCTAssertEqual(visibleFrame.y, 25, accuracy: 0.001)
-    }
-
-    @MainActor
-    func testLoadRepairsAndPersistsPoisonedAgentHookShellWrapperResumeBinding() throws {
-        let tempDir = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-session-tests-\(UUID().uuidString)", isDirectory: true)
-        try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: tempDir) }
-
-        let store = sessionStore()
-        let snapshotURL = tempDir.appendingPathComponent("session.json", isDirectory: false)
-        let source = Workspace()
-        let sourcePanelId = try XCTUnwrap(source.focusedPanelId)
-        let sessionId = "codex-load-repair-session"
-        let bindingIndex = SurfaceResumeBindingIndex(bindingsByPanel: [
-            SurfaceResumeBindingIndex.PanelKey(workspaceId: source.id, panelId: sourcePanelId): SurfaceResumeBindingSnapshot(
-                name: "Codex",
-                kind: "codex",
-                command: "{ cd -- '/tmp/right' 2>/dev/null || [ ! -d '/tmp/right' ]; } && 'bash' 'resume' '\(sessionId)'",
-                cwd: "/tmp/right",
-                checkpointId: sessionId,
-                source: "agent-hook",
-                autoResume: true,
-                updatedAt: 10
-            ),
-        ])
-        let workspaceSnapshot = source.sessionSnapshot(
-            includeScrollback: false,
-            surfaceResumeBindingIndex: bindingIndex
-        )
-        let appSnapshot = makeSnapshot(workspaceSnapshot: workspaceSnapshot)
-        XCTAssertTrue(store.save(appSnapshot, fileURL: snapshotURL))
-
-        let loaded = try XCTUnwrap(store.load(fileURL: snapshotURL))
-        let loadedBinding = loaded.windows.first?.tabManager.workspaces.first?.panels.first?.terminal?.resumeBinding
-        XCTAssertNil(loadedBinding)
-
-        let persistedData = try Data(contentsOf: snapshotURL)
-        let persisted = try JSONDecoder().decode(AppSessionSnapshot.self, from: persistedData)
-        let persistedBinding = persisted.windows.first?.tabManager.workspaces.first?.panels.first?.terminal?.resumeBinding
-        XCTAssertNil(persistedBinding)
-    }
-
-    @MainActor
-    func testLoadKeepsCustomShellNamedAgentHookResumeBinding() throws {
-        let tempDir = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-session-tests-\(UUID().uuidString)", isDirectory: true)
-        try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: tempDir) }
-
-        let store = sessionStore()
-        let snapshotURL = tempDir.appendingPathComponent("session.json", isDirectory: false)
-        let source = Workspace()
-        let sourcePanelId = try XCTUnwrap(source.focusedPanelId)
-        let command = "{ cd -- '/tmp/right' 2>/dev/null || [ ! -d '/tmp/right' ]; } && 'fish' 'resume' 'custom-session'"
-        let bindingIndex = SurfaceResumeBindingIndex(bindingsByPanel: [
-            SurfaceResumeBindingIndex.PanelKey(workspaceId: source.id, panelId: sourcePanelId): SurfaceResumeBindingSnapshot(
-                name: "Fish Agent",
-                kind: "fish",
-                command: command,
-                cwd: "/tmp/right",
-                checkpointId: "custom-session",
-                source: "agent-hook",
-                autoResume: true,
-                updatedAt: 10
-            ),
-        ])
-        let workspaceSnapshot = source.sessionSnapshot(
-            includeScrollback: false,
-            surfaceResumeBindingIndex: bindingIndex
-        )
-        let appSnapshot = makeSnapshot(workspaceSnapshot: workspaceSnapshot)
-        XCTAssertTrue(store.save(appSnapshot, fileURL: snapshotURL))
-
-        let loaded = try XCTUnwrap(store.load(fileURL: snapshotURL))
-        let loadedBinding = loaded.windows.first?.tabManager.workspaces.first?.panels.first?.terminal?.resumeBinding
-        XCTAssertEqual(loadedBinding?.command, command)
-    }
-
-    @MainActor
-    func testLoadRepairsWrongForkAgentLaunchCommandAndRecoversWorkingDirectory() throws {
-        let tempDir = FileManager.default.temporaryDirectory
-            .appendingPathComponent("cmux-session-tests-\(UUID().uuidString)", isDirectory: true)
-        try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: tempDir) }
-
-        let store = sessionStore()
-        let snapshotURL = tempDir.appendingPathComponent("session.json", isDirectory: false)
-        let source = Workspace()
-        let sourcePanelId = try XCTUnwrap(source.focusedPanelId)
-        source.updatePanelDirectory(panelId: sourcePanelId, directory: "/tmp/right")
-        var workspaceSnapshot = source.sessionSnapshot(includeScrollback: false)
-        let panelIndex = try XCTUnwrap(workspaceSnapshot.panels.firstIndex { $0.id == sourcePanelId })
-        var panel = workspaceSnapshot.panels[panelIndex]
-        var terminal = try XCTUnwrap(panel.terminal)
-        terminal.workingDirectory = "/tmp/right"
-        terminal.agent = SessionRestorableAgentSnapshot(
-            kind: .codex,
-            sessionId: "codex-load-repair-session",
-            workingDirectory: "/tmp/wrong",
-            launchCommand: AgentLaunchCommandSnapshot(
-                launcher: "claude",
-                executablePath: "/usr/local/bin/claude",
-                arguments: ["/usr/local/bin/claude", "--resume", "claude-session"],
-                workingDirectory: "/tmp/wrong",
-                environment: nil,
-                capturedAt: 123,
-                source: "process"
-            )
-        )
-        panel.directory = "/tmp/right"
-        panel.terminal = terminal
-        workspaceSnapshot.panels[panelIndex] = panel
-        let appSnapshot = makeSnapshot(workspaceSnapshot: workspaceSnapshot)
-        XCTAssertTrue(store.save(appSnapshot, fileURL: snapshotURL))
-
-        let loaded = try XCTUnwrap(store.load(fileURL: snapshotURL))
-        let loadedAgent = try XCTUnwrap(
-            loaded.windows.first?.tabManager.workspaces.first?.panels.first?.terminal?.agent
-        )
-        XCTAssertNil(loadedAgent.launchCommand)
-        XCTAssertEqual(loadedAgent.workingDirectory, "/tmp/right")
-        XCTAssertEqual(
-            loadedAgent.resumeCommand,
-            "{ cd -- '/tmp/right' 2>/dev/null || [ ! -d '/tmp/right' ]; } && 'codex' 'resume' 'codex-load-repair-session'"
-        )
     }
 
     func testLoadReopenSessionSnapshotRequiresPreviousSnapshotFile() throws {
@@ -1995,23 +1868,6 @@ final class SessionPersistenceTests: XCTestCase {
         )
     }
 
-    private func makeSnapshot(workspaceSnapshot: SessionWorkspaceSnapshot) -> AppSessionSnapshot {
-        let window = SessionWindowSnapshot(
-            frame: SessionRectSnapshot(x: 10, y: 20, width: 900, height: 700),
-            display: nil,
-            tabManager: SessionTabManagerSnapshot(
-                selectedWorkspaceIndex: 0,
-                workspaces: [workspaceSnapshot]
-            ),
-            sidebar: SessionSidebarSnapshot(isVisible: true, selection: .tabs, width: 240)
-        )
-        return AppSessionSnapshot(
-            version: SessionSnapshotSchema.currentVersion,
-            createdAt: Date().timeIntervalSince1970,
-            windows: [window]
-        )
-    }
-
     private func fileNumber(for fileURL: URL) throws -> Int {
         let attributes = try FileManager.default.attributesOfItem(atPath: fileURL.path)
         return try XCTUnwrap(attributes[.systemFileNumber] as? Int)
@@ -2053,78 +1909,6 @@ final class SocketListenerAcceptPolicyTests: XCTestCase {
         )
         // The captured real-binary path must not survive: it would bypass the wrapper.
         XCTAssertFalse(snapshot.resumeCommand?.contains("/opt/Claude Code/bin/claude") ?? true)
-    }
-
-    func testResumeCommandDropsMissingLauncherWrongForkLaunchCapture() {
-        let snapshot = SessionRestorableAgentSnapshot(
-            kind: .codex,
-            sessionId: "codex-session-123",
-            workingDirectory: "/tmp/repo",
-            launchCommand: AgentLaunchCommandSnapshot(
-                launcher: nil,
-                executablePath: "/usr/local/bin/claude",
-                arguments: ["/usr/local/bin/claude", "--resume", "claude-session"],
-                workingDirectory: "/tmp/wrong",
-                environment: ["CLAUDE_CONFIG_DIR": "/tmp/claude"],
-                capturedAt: 123,
-                source: "process"
-            )
-        )
-
-        XCTAssertEqual(
-            snapshot.resumeCommand,
-            "{ cd -- '/tmp/repo' 2>/dev/null || [ ! -d '/tmp/repo' ]; } && 'codex' 'resume' 'codex-session-123'"
-        )
-    }
-
-    func testResumeCommandPreservesMissingLauncherHermesLaunchCapture() {
-        let snapshot = SessionRestorableAgentSnapshot(
-            kind: .hermesAgent,
-            sessionId: "hermes-session-123",
-            workingDirectory: "/tmp/hermes repo",
-            launchCommand: AgentLaunchCommandSnapshot(
-                launcher: nil,
-                executablePath: "/opt/homebrew/bin/hermes",
-                arguments: [
-                    "/opt/homebrew/bin/hermes",
-                    "--provider",
-                    "custom",
-                    "--model",
-                    "gpt-5.5"
-                ],
-                workingDirectory: "/tmp/hermes repo",
-                environment: nil,
-                capturedAt: 123,
-                source: "process"
-            )
-        )
-
-        XCTAssertEqual(
-            snapshot.resumeCommand,
-            "{ cd -- '/tmp/hermes repo' 2>/dev/null || [ ! -d '/tmp/hermes repo' ]; } && '/opt/homebrew/bin/hermes' '--provider' 'custom' '--model' 'gpt-5.5' '--resume' 'hermes-session-123'"
-        )
-    }
-
-    func testResumeCommandPreservesMissingLauncherRovoDevAcliLaunchCapture() {
-        let snapshot = SessionRestorableAgentSnapshot(
-            kind: .rovodev,
-            sessionId: "rovo-session-123",
-            workingDirectory: "/tmp/rovo repo",
-            launchCommand: AgentLaunchCommandSnapshot(
-                launcher: nil,
-                executablePath: "/usr/local/bin/acli",
-                arguments: ["/usr/local/bin/acli", "rovodev", "run"],
-                workingDirectory: "/tmp/rovo repo",
-                environment: nil,
-                capturedAt: 123,
-                source: "process"
-            )
-        )
-
-        XCTAssertEqual(
-            snapshot.resumeCommand,
-            "{ cd -- '/tmp/rovo repo' 2>/dev/null || [ ! -d '/tmp/rovo repo' ]; } && '/usr/local/bin/acli' 'rovodev' 'run' '--restore' 'rovo-session-123'"
-        )
     }
 
     func testClaudeForkCommandRoutesThroughWrapperInsteadOfCapturedRealBinary() throws {
